[update]#34 背景画像の固定
- 既存の新規登録画面の削除
- 背景画像の差し替え
